refactor: 'Greeting' view to use new 'AbstractView' class; other modules to reference new 'Greetin'
